move component into new folder
