+++ frontend/src/scenes/insights/InsightNav/InsightsNav.tsx
@@ -1,8 +1,8 @@
 import { useActions, useValues } from 'kea'
 import { InsightType } from '~/types'
-import { insightLogic } from './insightLogic'
+import { insightLogic } from '../insightLogic'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
-import { FunnelsCue } from './views/Trends/FunnelsCue'
+import { FunnelsCue } from '../views/Trends/FunnelsCue'
 import { INSIGHT_TYPES_METADATA } from 'scenes/saved-insights/SavedInsights'
 import { Link } from 'lib/lemon-ui/Link'
 import { urls } from 'scenes/urls'
